refactor(p1): extract Sandbox composables Stage 2 — 14 composables (~673 LOC)
Move 14 Sandbox composables from ChatInput.kt (1472 → 783, -47%) to ChatInputSandbox.kt:

internal (called from outside the new file):
- SandboxPeekBar (main ChatInput in old file calls it)
- SandboxSheetHeader, SandboxToolActivityContent, SandboxWebActivityContent (Sheet still in old file calls them; will demote in Stage 3)

private (block-internal to new file):
- AgentOperationPreviewPeek, WebOperationPreviewThumbnail, HiddenWebOperationRenderer, WebOperationPreviewPlaceholder, SandboxStepPeek, SandboxStepArrow, SandboxStepStatusIcon, OperationWebPreview, WebOperationPreviewLoadingOverlay, SandboxSheetCodeBlock

Drop 16 imports from ChatInput.kt that became orphan via this move (AsyncImage, DisposableEffect, derivedStateOf, collectAsStateWithLifecycle, TransformOrigin, graphicsLayer, ContentScale, WebView, rememberWebViewState, WebViewLoadStatus, WebViewOperationState, WebViewOperationStore, ArrowLeft01, ArrowRight01, Code, Tick01). Preserve pre-existing dead WorkspaceTone import (CLAUDE.md §3).

Byte-equal composable bodies: only visibility modifier changed for the 4 promoted-to-internal ones. AnimatedVisibility / WebView state machine / Bitmap thumbnail capture / DisposableEffect lifecycle keys all preserved as-was.

Baseline 6/427 unit tests preserved. SandboxActivitySheet stays in ChatInput.kt — Stage 3 commit will move it (cross-package visibility change).
