+++ app/src/main/java/me/rerere/rikkahub/ui/components/ai/ChatInput.kt
@@ -50,11 +50,8 @@ import androidx.compose.material3.TextButton
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -67,11 +64,8 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.TransformOrigin
 import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.lerp
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
@@ -84,7 +78,6 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.util.fastForEach
 import androidx.core.content.FileProvider
 import androidx.core.net.toUri
-import coil3.compose.AsyncImage
 import com.dokar.sonner.ToastType
 import dev.chrisbanes.haze.HazeState
 import dev.chrisbanes.haze.hazeEffect
@@ -104,21 +97,14 @@ import me.rerere.ai.ui.UIMessagePart
 import me.rerere.common.android.appTempFolder
 import me.rerere.hugeicons.HugeIcons
 import me.rerere.hugeicons.stroke.Add01
-import me.rerere.hugeicons.stroke.ArrowLeft01
-import me.rerere.hugeicons.stroke.ArrowRight01
 import me.rerere.hugeicons.stroke.ArrowUp01
 import me.rerere.hugeicons.stroke.ArrowUp02
 import me.rerere.hugeicons.stroke.Book03
 import me.rerere.hugeicons.stroke.Cancel01
-import me.rerere.hugeicons.stroke.Code
-import me.rerere.hugeicons.stroke.Tick01
 import me.rerere.rikkahub.BuildConfig
 import me.rerere.rikkahub.R
 import me.rerere.rikkahub.data.agent.SandboxActivityUiState
 import me.rerere.rikkahub.data.agent.ToolActivityStatus
-import me.rerere.rikkahub.data.agent.webview.WebViewLoadStatus
-import me.rerere.rikkahub.data.agent.webview.WebViewOperationState
-import me.rerere.rikkahub.data.agent.webview.WebViewOperationStore
 import me.rerere.rikkahub.data.ai.vision.ImageAttachmentValidator
 import me.rerere.rikkahub.data.datastore.Settings
 import me.rerere.rikkahub.data.datastore.findProvider
@@ -132,8 +118,6 @@ import me.rerere.rikkahub.service.PendingUserMessageMode
 import me.rerere.rikkahub.ui.components.ui.KeepScreenOn
 import me.rerere.rikkahub.ui.components.ui.WorkspaceIconButton
 import me.rerere.rikkahub.ui.components.ui.WorkspaceTone
-import me.rerere.rikkahub.ui.components.webview.WebView
-import me.rerere.rikkahub.ui.components.webview.rememberWebViewState
 import me.rerere.rikkahub.ui.context.LocalSettings
 import me.rerere.rikkahub.ui.context.LocalToaster
 import me.rerere.rikkahub.ui.components.ui.workspaceColors
@@ -718,362 +702,6 @@ fun ChatInput(
 }
 
 @Composable
-private fun SandboxPeekBar(
-    activity: SandboxActivityUiState,
-    onOpen: () -> Unit,
-    onCancel: (() -> Unit)?,
-    onPrevious: (() -> Unit)?,
-    onNext: (() -> Unit)?,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .animateContentSize(),
-        verticalAlignment = Alignment.Bottom,
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        AgentOperationPreviewPeek(
-            activity = activity,
-            onOpen = onOpen,
-            modifier = Modifier
-                .width(118.dp)
-                .height(78.dp),
-        )
-        SandboxStepPeek(
-            activity = activity,
-            onOpen = onOpen,
-            onCancel = onCancel,
-            onPrevious = onPrevious,
-            onNext = onNext,
-            modifier = Modifier
-                .weight(1f)
-                .height(38.dp),
-        )
-    }
-}
-
-@Composable
-private fun AgentOperationPreviewPeek(
-    activity: SandboxActivityUiState,
-    onOpen: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val previewUrl = activity.operationPreviewUrl()
-    val workspace = workspaceColors()
-    Surface(
-        modifier = modifier
-            .clip(RoundedCornerShape(10.dp))
-            .clickable { onOpen() },
-        shape = RoundedCornerShape(10.dp),
-        color = workspace.paper,
-        contentColor = workspace.ink,
-        shadowElevation = 1.dp,
-        border = BorderStroke(1.dp, workspace.hairline),
-    ) {
-        if (previewUrl != null) {
-            WebOperationPreviewThumbnail(
-                url = previewUrl,
-                toolCallId = activity.toolCallId,
-                onOpen = onOpen,
-            )
-        } else {
-            Column(
-                modifier = Modifier.padding(horizontal = 9.dp, vertical = 7.dp),
-                verticalArrangement = Arrangement.spacedBy(2.dp),
-            ) {
-                Text(
-                    text = activity.operationPreviewKind(),
-                    color = workspace.amber,
-                    style = MaterialTheme.typography.labelSmall,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-                Text(
-                    text = activity.operationPreviewText(),
-                    color = workspace.muted,
-                    style = MaterialTheme.typography.labelSmall.copy(fontFamily = FontFamily.Monospace),
-                    maxLines = 4,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun WebOperationPreviewThumbnail(
-    url: String,
-    toolCallId: String,
-    onOpen: () -> Unit,
-) {
-    val webViewOperationStore: WebViewOperationStore = koinInject()
-    val webState by webViewOperationStore.state.collectAsStateWithLifecycle()
-    val normalizedUrl = remember(url) { url.normalizedWebPreviewUrl() }
-    // Key only on toolCallId + normalizedUrl (stable), not on webState fields (change on every capture)
-    val isCurrentPreview = remember(toolCallId, normalizedUrl, webState.loadId) {
-        webState.matchesPreview(toolCallId = toolCallId, normalizedUrl = normalizedUrl)
-    }
-    val thumbnailFile = webState.bestThumbnailFile(isCurrentPreview)
-    LaunchedEffect(toolCallId, normalizedUrl, isCurrentPreview) {
-        if (!isCurrentPreview) {
-            webViewOperationStore.open(url, toolCallId = toolCallId)
-        }
-    }
-    val loadId = webState.loadId.takeIf { isCurrentPreview }
-    val workspace = workspaceColors()
-    Box(modifier = Modifier.fillMaxSize()) {
-        if (!loadId.isNullOrBlank()) {
-            HiddenWebOperationRenderer(
-                url = url,
-                loadId = loadId,
-                webViewOperationStore = webViewOperationStore,
-                modifier = Modifier
-                    .requiredSize(width = 336.dp, height = 252.dp)
-                    .graphicsLayer {
-                        alpha = 0f
-                        scaleX = 0.35f
-                        scaleY = 0.35f
-                        transformOrigin = TransformOrigin(0f, 0f)
-                    },
-            )
-        }
-        if (thumbnailFile != null) {
-            AsyncImage(
-                model = thumbnailFile,
-                contentDescription = null,
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop,
-            )
-        } else {
-            WebOperationPreviewPlaceholder(
-                url = url,
-                webState = webState.takeIf { isCurrentPreview },
-                workspace = workspace,
-            )
-        }
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .clickable { onOpen() },
-        )
-    }
-}
-
-@Composable
-private fun HiddenWebOperationRenderer(
-    url: String,
-    loadId: String,
-    webViewOperationStore: WebViewOperationStore,
-    modifier: Modifier = Modifier,
-) {
-    val state = rememberWebViewState(
-        url = url,
-        settings = {
-            useWideViewPort = true
-            loadWithOverviewMode = true
-            textZoom = 85
-            disablePreviewDarkening()
-        },
-    )
-    DisposableEffect(loadId) {
-        webViewOperationStore.markRendererActive(loadId, true)
-        onDispose {
-            webViewOperationStore.markRendererActive(loadId, false)
-        }
-    }
-    WebView(
-        state = state,
-        modifier = modifier,
-        onCreated = { webView ->
-            webView.isFocusable = false
-            webView.isFocusableInTouchMode = false
-            webView.setOnTouchListener { _, _ -> true }
-            webViewOperationStore.markRendererActive(loadId, true)
-        },
-        onUpdated = { webView ->
-            webView.setOnTouchListener { _, _ -> true }
-        },
-        onProgressChanged = { webView, progress ->
-            webViewOperationStore.updateLoading(loadId, webView?.url ?: url, progress)
-            if (progress >= 35) {
-                webView?.let { extractReadablePage(it, webViewOperationStore, loadId) }
-            }
-        },
-        onPageStarted = { webView, pageUrl ->
-            webViewOperationStore.updateLoading(loadId, pageUrl ?: webView?.url ?: url, 1)
-            webView?.let { scheduleReadableExtracts(it, webViewOperationStore, loadId) }
-        },
-        onPageFinished = { webView, pageUrl ->
-            val resolvedUrl = pageUrl ?: webView?.url ?: url
-            webViewOperationStore.markPageFinished(loadId, resolvedUrl)
-            webView?.let {
-                extractReadablePage(it, webViewOperationStore, loadId, force = true)
-                scheduleThumbnailCaptures(it, webViewOperationStore, context = it.context, loadId = loadId)
-            }
-        },
-        onReceivedError = { webView, pageUrl, error ->
-            webViewOperationStore.markFailed(loadId, pageUrl ?: webView?.url ?: url, error.orEmpty())
-        },
-    )
-}
-
-@Composable
-private fun WebOperationPreviewPlaceholder(
-    url: String,
-    webState: WebViewOperationState?,
-    workspace: me.rerere.rikkahub.ui.components.ui.WorkspaceColors,
-) {
-    val title = webState?.title.orEmpty().ifBlank { "网页预览" }
-    val detail = when {
-        webState?.lastError?.isNotBlank() == true -> webState.lastError
-        webState?.isLoading == true -> "正在加载 ${url.webHostPreview()}"
-        webState?.status == WebViewLoadStatus.STALLED -> "网页加载较慢"
-        else -> url
-    }
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(workspace.paper),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.labelMedium,
-            color = workspace.ink,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-        )
-        Text(
-            text = detail,
-            style = MaterialTheme.typography.labelSmall,
-            color = workspace.muted,
-            maxLines = 2,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.padding(horizontal = 8.dp),
-        )
-    }
-}
-
-@Composable
-private fun SandboxStepPeek(
-    activity: SandboxActivityUiState,
-    onOpen: () -> Unit,
-    onCancel: (() -> Unit)?,
-    onPrevious: (() -> Unit)?,
-    onNext: (() -> Unit)?,
-    modifier: Modifier = Modifier,
-) {
-    val workspace = workspaceColors()
-    Surface(
-        modifier = modifier
-            .clip(RoundedCornerShape(10.dp))
-            .clickable { onOpen() },
-        shape = RoundedCornerShape(10.dp),
-        color = workspace.paper,
-        contentColor = workspace.ink,
-        shadowElevation = 1.dp,
-        tonalElevation = 0.dp,
-        border = BorderStroke(1.dp, workspace.hairline),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 8.dp, vertical = 4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(6.dp),
-        ) {
-            SandboxStepStatusIcon(status = activity.status)
-            Text(
-                text = activity.title,
-                modifier = Modifier.weight(1f),
-                style = MaterialTheme.typography.labelMedium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-            if (onCancel != null) {
-                IconButton(
-                    onClick = onCancel,
-                    modifier = Modifier.size(22.dp),
-                ) {
-                    Icon(
-                        imageVector = HugeIcons.Cancel01,
-                        contentDescription = stringResource(R.string.stop),
-                        tint = Color(0xFFE09A1B),
-                        modifier = Modifier.size(12.dp),
-                    )
-                }
-            }
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(2.dp),
-            ) {
-                SandboxStepArrow(
-                    enabled = onPrevious != null,
-                    onClick = onPrevious,
-                    left = true,
-                )
-                Text(
-                    text = activity.stepProgressText(),
-                    style = MaterialTheme.typography.labelMedium,
-                    color = workspace.muted,
-                    maxLines = 1,
-                )
-                SandboxStepArrow(
-                    enabled = onNext != null,
-                    onClick = onNext,
-                    left = false,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SandboxStepArrow(
-    enabled: Boolean,
-    onClick: (() -> Unit)?,
-    left: Boolean,
-) {
-    Box(
-        modifier = Modifier
-            .size(20.dp)
-            .clip(CircleShape)
-            .clickable(enabled = enabled && onClick != null) { onClick?.invoke() },
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            imageVector = if (left) HugeIcons.ArrowLeft01 else HugeIcons.ArrowRight01,
-            contentDescription = null,
-            modifier = Modifier.size(12.dp),
-            tint = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = if (enabled) 0.92f else 0.28f),
-        )
-    }
-}
-
-@Composable
-private fun SandboxStepStatusIcon(status: ToolActivityStatus) {
-    Surface(
-        modifier = Modifier.size(20.dp),
-        shape = CircleShape,
-        color = sandboxStatusContainerColor(status),
-        contentColor = sandboxStatusOnContainerColor(status),
-    ) {
-        Box(contentAlignment = Alignment.Center) {
-            Icon(
-                imageVector = when (status) {
-                    ToolActivityStatus.SUCCEEDED -> HugeIcons.Tick01
-                    ToolActivityStatus.FAILED,
-                    ToolActivityStatus.CANCELLED -> HugeIcons.Cancel01
-                    else -> HugeIcons.Code
-                },
-                contentDescription = null,
-                modifier = Modifier.size(12.dp),
-            )
-        }
-    }
-}
-
-@Composable
 fun SandboxActivitySheet(
     activity: SandboxActivityUiState,
     onDismiss: () -> Unit,
@@ -1120,323 +748,6 @@ fun SandboxActivitySheet(
                         .weight(1f),
                 )
             }
-        }
-    }
-}
-
-@Composable
-private fun SandboxSheetHeader(
-    activity: SandboxActivityUiState,
-    isWebPreview: Boolean,
-    onCancel: (() -> Unit)?,
-    onPrevious: (() -> Unit)?,
-    onNext: (() -> Unit)?,
-) {
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(22.dp),
-        color = MaterialTheme.colorScheme.surface,
-        contentColor = MaterialTheme.colorScheme.onSurface,
-        tonalElevation = 1.dp,
-        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.75f)),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 9.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            SandboxStepStatusIcon(status = activity.status)
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = activity.title,
-                    style = MaterialTheme.typography.titleSmall,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-                val supporting = when {
-                    isWebPreview -> activity.operationPreviewUrl()?.webHostPreview().orEmpty()
-                    activity.runtime.isNotBlank() -> activity.runtime
-                    else -> activity.toolName
-                }
-                Text(
-                    text = supporting,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-            if (activity.canCancel && onCancel != null) {
-                Surface(
-                    onClick = onCancel,
-                    shape = CircleShape,
-                    color = MaterialTheme.colorScheme.tertiaryContainer,
-                    contentColor = MaterialTheme.colorScheme.onTertiaryContainer,
-                ) {
-                    Text(
-                        text = "中断",
-                        modifier = Modifier.padding(horizontal = 10.dp, vertical = 6.dp),
-                        style = MaterialTheme.typography.labelSmall,
-                        maxLines = 1,
-                    )
-                }
-            }
-            SandboxStepArrow(
-                enabled = onPrevious != null,
-                onClick = onPrevious,
-                left = true,
-            )
-            Text(
-                text = activity.stepProgressText(),
-                style = MaterialTheme.typography.titleSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                maxLines = 1,
-            )
-            SandboxStepArrow(
-                enabled = onNext != null,
-                onClick = onNext,
-                left = false,
-            )
-        }
-    }
-}
-
-@Composable
-private fun SandboxToolActivityContent(
-    activity: SandboxActivityUiState,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier.verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        SandboxSheetCodeBlock(
-            title = "调用内容",
-            language = if (activity.toolName.startsWith("terminal_")) "shell" else "text",
-            content = activity.inputPreview.ifBlank { activity.title },
-        )
-        SandboxSheetCodeBlock(
-            title = "调用结果",
-            language = "text",
-            content = activity.outputTail.ifBlank {
-                if (activity.status == ToolActivityStatus.RUNNING || activity.status == ToolActivityStatus.WAITING_FOR_PERMISSION) {
-                    "等待工具返回输出..."
-                } else {
-                    "无输出"
-                }
-            },
-        )
-    }
-}
-
-@Composable
-private fun SandboxWebActivityContent(
-    activity: SandboxActivityUiState,
-    url: String,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        Surface(
-            modifier = Modifier.fillMaxWidth(),
-            shape = CircleShape,
-            color = MaterialTheme.colorScheme.surfaceContainerHighest,
-            contentColor = MaterialTheme.colorScheme.onSurfaceVariant,
-        ) {
-            Text(
-                text = url,
-                modifier = Modifier.padding(horizontal = 12.dp, vertical = 7.dp),
-                style = MaterialTheme.typography.labelMedium,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-        OperationWebPreview(
-            url = url,
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f),
-        )
-    }
-}
-
-@Composable
-private fun OperationWebPreview(
-    url: String,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-    val webViewOperationStore: WebViewOperationStore = koinInject()
-    val webState by webViewOperationStore.state.collectAsStateWithLifecycle()
-    val normalizedUrl = remember(url) { url.normalizedWebPreviewUrl() }
-    // Key only on normalizedUrl + loadId (stable), not on all webState fields
-    val isCurrentPreview = remember(normalizedUrl, webState.loadId) {
-        webState.matchesPreview(toolCallId = "", normalizedUrl = normalizedUrl)
-    }
-    LaunchedEffect(normalizedUrl, isCurrentPreview) {
-        if (!isCurrentPreview) {
-            webViewOperationStore.open(url)
-        }
-    }
-    val loadId = webState.loadId.takeIf { isCurrentPreview }
-    DisposableEffect(loadId) {
-        if (!loadId.isNullOrBlank()) {
-            webViewOperationStore.markRendererActive(loadId, true)
-        }
-        onDispose {
-            if (!loadId.isNullOrBlank()) {
-                webViewOperationStore.markRendererActive(loadId, false)
-            }
-        }
-    }
-    val thumbnailFile = webState.bestThumbnailFile(isCurrentPreview)
-    val state = rememberWebViewState(url = url)
-    // derivedStateOf keyed on loadId so status transitions are coalesced within a single load.
-    // loadId changes rarely (only on new navigation), so the derivedStateOf instance is stable.
-    val statusReady by remember(loadId) {
-        derivedStateOf {
-            webState.status in setOf(WebViewLoadStatus.INTERACTIVE, WebViewLoadStatus.READY)
-        }
-    }
-    val showLiveWebView = !loadId.isNullOrBlank() && statusReady
-    Surface(
-        modifier = modifier,
-        shape = RoundedCornerShape(18.dp),
-        color = MaterialTheme.colorScheme.surface,
-        contentColor = MaterialTheme.colorScheme.onSurface,
-        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant),
-    ) {
-        Box(modifier = Modifier.fillMaxSize()) {
-            if (!loadId.isNullOrBlank()) {
-                WebView(
-                    state = state,
-                    modifier = Modifier.fillMaxSize(),
-                    onProgressChanged = { webView, progress ->
-                        webViewOperationStore.updateLoading(loadId, webView?.url ?: url, progress)
-                        if (progress >= 35) {
-                            webView?.let { view -> extractReadablePage(view, webViewOperationStore, loadId) }
-                        }
-                    },
-                    onPageStarted = { webView, pageUrl ->
-                        webViewOperationStore.updateLoading(loadId, pageUrl ?: webView?.url ?: url, 1)
-                        webView?.let { view -> scheduleReadableExtracts(view, webViewOperationStore, loadId) }
-                    },
-                    onPageFinished = { webView, pageUrl ->
-                        val resolvedUrl = pageUrl ?: webView?.url ?: url
-                        webViewOperationStore.markPageFinished(loadId, resolvedUrl)
-                        webView?.let { view ->
-                            extractReadablePage(view, webViewOperationStore, loadId, force = true)
-                            scheduleThumbnailCaptures(view, webViewOperationStore, context, loadId)
-                        }
-                    },
-                    onReceivedError = { webView, pageUrl, error ->
-                        webViewOperationStore.markFailed(loadId, pageUrl ?: webView?.url ?: url, error.orEmpty())
-                    },
-                )
-            }
-            if (!showLiveWebView) {
-                WebOperationPreviewLoadingOverlay(
-                    url = url,
-                    webState = webState.takeIf { isCurrentPreview },
-                    thumbnailFile = thumbnailFile,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun WebOperationPreviewLoadingOverlay(
-    url: String,
-    webState: WebViewOperationState?,
-    thumbnailFile: File?,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.surface),
-    ) {
-        if (thumbnailFile != null) {
-            AsyncImage(
-                model = thumbnailFile,
-                contentDescription = null,
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop,
-            )
-        } else {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(18.dp),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                Text(
-                    text = webState?.title?.takeIf { it.isNotBlank() } ?: "正在加载网页",
-                    style = MaterialTheme.typography.titleSmall,
-                    color = MaterialTheme.colorScheme.onSurface,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-                Spacer(modifier = Modifier.height(6.dp))
-                Text(
-                    text = when {
-                        webState?.lastError?.isNotBlank() == true -> webState.lastError
-                        webState?.readableText?.isNotBlank() == true -> webState.readableText.compactForSandbox(120)
-                        webState?.status == WebViewLoadStatus.STALLED -> "网页加载较慢，正在保留当前预览"
-                        else -> url.webHostPreview()
-                    },
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    maxLines = 3,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SandboxSheetCodeBlock(
-    title: String,
-    language: String,
-    content: String,
-) {
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(18.dp),
-        color = MaterialTheme.colorScheme.surface,
-        contentColor = MaterialTheme.colorScheme.onSurface,
-        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outlineVariant),
-    ) {
-        Column {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.surfaceContainerHighest)
-                    .padding(horizontal = 14.dp, vertical = 9.dp),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween,
-            ) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.labelLarge,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                Text(
-                    text = language,
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.72f),
-                )
-            }
-            Text(
-                text = content,
-                modifier = Modifier.padding(horizontal = 14.dp, vertical = 12.dp),
-                style = MaterialTheme.typography.bodyMedium.copy(fontFamily = FontFamily.Monospace),
-                color = MaterialTheme.colorScheme.onSurface,
-            )
         }
     }
 }
